+++ config/default.rb
@@ -11,6 +11,7 @@ Vagrant::Config.run do |config|
   config.ssh.timeout = 10
   config.ssh.forward_agent = false
   config.ssh.forward_x11 = false
+  config.ssh.shell = "bash"
 
   config.vm.auto_port_range = (2200..2250)
   config.vm.box_url = nil