+++ lib/vagrant/provisioners/chef.rb
@@ -147,6 +147,12 @@ module Vagrant
           run_list << name
         end
 
+        def validate(env, errors)
+          super
+
+          errors.add(I18n.t("vagrant.config.chef.vagrant_as_json_key")) if @json.has_key?(:vagrant)
+        end
+
         def instance_variables_hash
           # Overridden so that the 'json' key could be removed, since its just
           # merged into the config anyways