+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.9.2 (unreleased)
 
-
+  - Fix `forward_agent` not working when outside of blocks. [GH-651]
 
 ## 0.9.1 (January 18, 2012)