+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
   - Fix `forward_agent` not working when outside of blocks. [GH-651]
   - Fix issue causing custom guest implementations to not load properly.
+  - Filter clear screen character out of output on SSH.
 
 ## 0.9.1 (January 18, 2012)